requirements.txt
flask==3.0.3
requests==2.31.0
google-generativeai==0.5.2
python-dotenv==1.0.1